Fix for Identifier Kung-Fu
While working on my own data, I saw that I was missing important genes after performing the identifier transformation. I checked and in the Uniprot file there are many Ensembl IDs missing (only ~11k when in reality there are ~67k). This is a problem because the users end up losing genes that can be significant. Moreover, sometimes different Ensembl IDs map to the same gene symbol, because we only want one measure of expression per gene I added the code to compute the mean between them.